Split mantle into transactions/ops, rename encoding into codec
