+++ core/benches/mantle_tx_components.rs
@@ -18,7 +18,6 @@ use logos_blockchain_core::{
     crypto::{Hasher, ZkHasher},
     mantle::{
         MantleTx, SignedMantleTx, Transaction as _, TxHash,
-        encoding::{decode_signed_mantle_tx, encode_mantle_tx, encode_signed_mantle_tx},
         ops::{
             Op, OpProof,
             channel::{
@@ -26,6 +25,7 @@ use logos_blockchain_core::{
                 inscribe::{Inscription, InscriptionOp},
             },
         },
+        transactions::codec::{decode_signed_mantle_tx, encode_mantle_tx, encode_signed_mantle_tx},
     },
 };
