+++ core/src/mantle/transactions/builder.rs
@@ -4,15 +4,13 @@ use lb_key_management_system_keys::keys::ZkPublicKey;
 use lb_utils::bounded_vec::BoundedError;
 use thiserror::Error;
 
-use super::{GasCalculator as _, GasConstants, MantleTx, Note, Op, Utxo};
 use crate::{
     mantle::{
-        NoteId,
-        encoding::BoundedUtxos,
+        GasCalculator as _, GasConstants, Note, NoteId, Op, Utxo,
         gas::{GasCost, GasOverflow},
-        ledger::{Inputs, Outputs},
+        ledger::{BoundedUtxos, Inputs, Outputs},
         ops::{channel::withdraw::ChannelWithdrawOp, transfer::TransferOp},
-        tx::{GasPrices, MantleTxContext},
+        transactions::{GasPrices, MantleTx, MantleTxContext},
     },
     proofs::channel_multi_sig_proof::ChannelMultiSigProof,
 };
@@ -275,7 +273,7 @@ mod tests {
                 leader_claim::LeaderClaimOp,
                 sdp::{SDPDeclareOp, SDPWithdrawOp},
             },
-            tx::MantleTxGasContext,
+            transactions::MantleTxGasContext,
         },
         sdp::{DeclarationId, Locator, ProviderId, ServiceType},
     };
