+++ core/src/mantle/transactions/genesis_tx.rs
@@ -3,20 +3,24 @@ use nom::IResult;
 use serde::{Deserialize, Serialize};
 use time::OffsetDateTime;
 
-use super::{OpProof, SignedMantleTx, ops::sdp::SDPDeclareOp};
+use super::{SignedMantleTx, TxHash};
 use crate::{
     crypto::{Digest as _, Hasher},
     mantle::{
-        MantleTx, Transaction, TransactionHasher, TxHash,
-        encoding::{
-            decode_field_element, decode_uint64, decode_unix_timestamp, decode_utf8_string,
-            encode_field_element, encode_string, encode_uint64, encode_unix_timestamp,
-            proof_matches,
+        MantleTx, OpProof, Transaction, TransactionHasher,
+        codec::{
+            crypto::{decode_field_element, encode_field_element},
+            primitives::{
+                decode_uint64, decode_unix_timestamp, decode_utf8_string, encode_string,
+                encode_uint64, encode_unix_timestamp,
+            },
         },
         gas::{Gas, GasCalculator, GasConstants, GasCost, GasOverflow, GasPrice},
         ops::{
             Op,
             channel::{ChannelId, MsgId, inscribe::InscriptionOp},
+            codec::proof_matches,
+            sdp::SDPDeclareOp,
             transfer::TransferOp,
         },
     },
@@ -200,19 +204,19 @@ impl GasCalculator for GenesisTx {
 }
 
 impl crate::mantle::GenesisTx for GenesisTx {
-    fn genesis_inscription(&self) -> &InscriptionOp {
-        // Safe to unwrap because we validated this in from_tx
-        match &self.mantle_tx().ops()[1] {
-            Op::ChannelInscribe(op) => op,
-            _ => unreachable!("GenesisTx always has a valid inscription as second op"),
-        }
-    }
-
     fn genesis_transfer(&self) -> &TransferOp {
         // Safe to unwrap because we validated this in from_tx
         match &self.mantle_tx().ops()[0] {
             Op::Transfer(op) => op,
             _ => unreachable!("GenesisTx always has a valid transfer as first op"),
+        }
+    }
+
+    fn genesis_inscription(&self) -> &InscriptionOp {
+        // Safe to unwrap because we validated this in from_tx
+        match &self.mantle_tx().ops()[1] {
+            Op::ChannelInscribe(op) => op,
+            _ => unreachable!("GenesisTx always has a valid inscription as second op"),
         }
     }
 
@@ -326,9 +330,9 @@ mod tests {
     use super::*;
     use crate::{
         mantle::{
-            encoding::Ops,
             ledger::{Inputs, Note, Outputs, Utxo, Value},
             ops::channel::{Ed25519PublicKey, inscribe::Inscription},
+            transactions::Ops,
         },
         sdp::{Locator, ProviderId, ServiceType},
     };
