+++ ledger/src/cryptarchia/mod.rs
@@ -10,9 +10,9 @@ use lb_core::{
     mantle::{
         GenesisTx, NoteId, TxHash, Utxo, Value,
         gas::{Gas, GasConstants, GasCost, GasPrice},
-        genesis_tx::{GENESIS_EXECUTION_GAS_PRICE, GENESIS_STORAGE_GAS_PRICE},
         ledger::Operation as _,
         ops::transfer::{TransferOp, TransferValidationContext},
+        transactions::{GENESIS_EXECUTION_GAS_PRICE, GENESIS_STORAGE_GAS_PRICE},
     },
     proofs::leader_proof::{self, LeaderPublic},
     sdp::{Declarations, locked_notes::LockedNotes},
@@ -750,7 +750,7 @@ pub mod tests {
             gas::MainnetGasConstants,
             ledger::{Inputs, Outputs},
             ops::{leader_claim::VoucherCm, sdp::SDPDeclareOp},
-            tx::GasPrices,
+            transactions::GasPrices,
         },
         sdp::{Declaration, DeclarationId, Locator, ServiceParameters, ServiceType},
     };
