+++ tests/src/cucumber/error.rs
@@ -1,7 +1,7 @@
 use std::io::Error as IoError;
 
 use hex::FromHexError;
-use lb_core::{codec::Error, mantle::tx::VerificationError};
+use lb_core::{codec::Error, mantle::transactions::VerificationError};
 use lb_testing_framework::configs::wallet::WalletConfigError;
 use lb_wallet::WalletError;
 use lb_zksign::ZkSignError;
