+++ tools/config/src/consensus.rs
@@ -5,8 +5,6 @@ use lb_core::{
     block::genesis::{GenesisBlock, GenesisBlockBuilder},
     mantle::{
         CryptarchiaParameter, MantleTx, Note, NoteId, OpProof, Utxo,
-        encoding::Ops,
-        genesis_tx::GenesisTx,
         ops::{
             Op, OpId as _,
             channel::{
@@ -15,6 +13,7 @@ use lb_core::{
             },
             transfer::TransferOp,
         },
+        transactions::{GenesisTx, Ops},
     },
     sdp::{DeclarationMessage, Locator, ProviderId, ServiceType},
 };
