+++ wallet/src/error.rs
@@ -3,7 +3,7 @@ use lb_core::{
     mantle::{
         gas::GasOverflow,
         ledger::{InputsError, OutputsError},
-        tx_builder::TxBuilderError,
+        transactions::builder::TxBuilderError,
     },
 };
 use lb_utils::bounded_vec::BoundedError;
